Move serializeIR and friends to ir_serializer
